Moved testReadQuantizedVectorWithEmptyRawVectors test case to BaserKnnVectorsFormatTestCase
