+++ lucene/core/src/test/org/apache/lucene/codecs/lucene99/TestLucene99HnswVectorsFormat.java
@@ -95,4 +95,9 @@ public class TestLucene99HnswVectorsFormat extends BaseKnnVectorsFormatTestCase 
       }
     }
   }
+
+  @Override
+  protected boolean isScalarQuantizedVectorsFormat() {
+    return false;
+  }
 }
