+++ lucene/sandbox/src/test/org/apache/lucene/sandbox/codecs/faiss/TestFaissKnnVectorsFormat.java
@@ -138,4 +138,9 @@ public class TestFaissKnnVectorsFormat extends BaseKnnVectorsFormatTestCase {
           Collections.nCopies(numDocs, List.of(new KnnFloatVectorField("vector", largeVector))));
     }
   }
+
+  @Override
+  protected boolean isScalarQuantizedVectorsFormat() {
+    return false;
+  }
 }
